requirements.txt
ariadne==0.21
black==23.10.1
fastapi==0.108.0
flake8==6.1.0
httpx==0.26.0
psycopg2-binary==2.9.9
pydantic==2.5.3
python-multipart==0.0.6
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
python-jose==3.3.0
SQLAlchemy==2.0.24
uvicorn==0.25.0

